image-slicer/requirements.txt
Pillow == 10.1.0
black == 23.11.0
python-dotenv == 1.0.0